db: cast active_quota_grants SUM back to bigint
`SUM(amount)` over a BIGINT column returns NUMERIC in postgres, but
`query_scalar` decodes into i64 — so once any quota grant existed,
every quota-checked MCP tool call failed with "mismatched types; Rust
type i64 (INT8) is not compatible with SQL type NUMERIC" (column 0).
This blocked all agents on every quota-consuming endpoint (get_tweet,
reply, post, like, ...) while only read_queue survived. Cast the sum
to ::bigint so it decodes cleanly.

+++ psychological-operations-db/src/quota.rs
@@ -84,8 +84,12 @@ impl Db {
         direction: &str,
         now: i64,
     ) -> Result<i64, Error> {
+        // `SUM(bigint)` yields NUMERIC in postgres; cast back to BIGINT so
+        // `query_scalar` can decode it as `i64` (otherwise: "mismatched
+        // types; Rust type i64 (INT8) is not compatible with SQL type
+        // NUMERIC" on every quota-checked tool call once any grant exists).
         let total: i64 = sqlx::query_scalar(
-            "SELECT COALESCE(SUM(amount), 0) FROM quota_grants \
+            "SELECT COALESCE(SUM(amount), 0)::bigint FROM quota_grants \
              WHERE account = $1 AND direction = $2 \
                AND granted_at <= $3 AND expires_at > $3",
         )
